backend/requirements.txt
asgiref==3.8.1
Django==5.1.0
django-cors-headers==4.6.0
django-filter==25.1
django-mongodb-backend==5.1.0b1
django-rest-knox==5.0.2
djangorestframework==3.16.0
dnspython==2.7.0
pillow==11.2.1
pymongo==4.11
python-decouple==3.8
sqlparse==0.5.3
